refactor: extract path-scoped config store
