+++ test/install_setup.test.ts
@@ -109,7 +109,7 @@ describe('install setup', () => {
 
   const installConfigSources = () => {
     fs.mkdirSync(path.join(repoDir, 'config'), { recursive: true });
-    ['.defaultConfig.json', 'index.ts', 'updateConfig.ts'].forEach((fileName) => {
+    ['.defaultConfig.json', 'index.ts', 'store.ts', 'updateConfig.ts'].forEach((fileName) => {
       fs.copyFileSync(
         path.join(repoRoot, 'config', fileName),
         path.join(repoDir, 'config', fileName),
